do not clobber existing /etc/raspindi.conf

+++ install.sh
@@ -17,7 +17,9 @@ fi
 cp build/src/raspindi "$BIN_DIR"
 cp build/src/libndioutput.so "$LIB_DIR"
 
-cp etc/raspindi.conf.default /etc/raspindi.conf
+if [ ! -e /etc/raspindi.conf ]; then
+  cp etc/raspindi.conf.default /etc/raspindi.conf
+fi
 
 echo '#!/usr/bin/env sh
 LD_LIBRARY_PATH="/opt/raspindi/lib" /opt/raspindi/bin/raspindi
